requirements.txt
Flask
flask-cors
Flask-JWT-Extended
Flask-Mail
Flask-Migrate
Flask-SQLAlchemy
google-auth
google-genai
neo4j-driver
pre_commit
psycopg2-binary
python-dotenv
redis
ruff
